Fix local stack's dynamo db init
